# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid input
        type: string
    type: object
...
  handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
  handlers.TaskResponse:
    properties:
      assignee:
        $ref: '#/definitions/handlers.UserResponse'
      assignee_id:
        example: 2
        type: integer
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        example: Implement the new feature
        type: string
      id:
        example: 1
        type: integer
      project_id:
        example: 1
        type: integer
      status:
        example: pending
        type: string
      title:
        example: Complete feature
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  handlers.UserResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      last_login_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      name:
        example: John Doe
        type: string
      two_factor_enabled:
        example: false
        type: boolean
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  services.CreateTaskInput:
    properties:
      assignee_id:
        description: Optional assignee
        type: integer
      description:
        description: Optional description
        type: string
      title:
        description: Task title is required
        type: string
    required:
    - title
    type: object
...
  services.UpdateTaskInput:
    properties:
      assignee_id:
        description: Optional assignee update
        type: integer
      description:
        description: Optional description update
        type: string
      status:
        description: Optional status update (pending, in_progress, completed)
        type: string
      title:
        description: Optional title update
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive task management API with real-time collaboration features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Real Task API
  version: "1.0"
paths:
...
  /api/projects/{project_id}/tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks within a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks retrieved successfully
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks for a project
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task within a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Task details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateTaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /api/tasks/{task_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing task
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid task ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Get details of a specific task by ID
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task retrieved successfully
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Invalid task ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a specific task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update an existing task's details
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: Updated task details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Tasks
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"